@@ -14,18 +14,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.3",
-    "@angular/cli": "1.6.7",
-    "@angular/common": "^5.2.3",
-    "@angular/compiler": "^5.2.0",
-    "@angular/compiler-cli": "^5.2.3",
-    "@angular/core": "^5.2.3",
-    "@angular/forms": "^5.2.3",
-    "@angular/http": "^5.2.3",
-    "@angular/platform-browser": "^5.2.3",
-    "@angular/platform-browser-dynamic": "^5.2.3",
-    "@angular/router": "^5.2.3",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0",
+    "@angular/animations": "6.1.6",
+    "@angular/common": "^6.1.6",
+    "@angular/compiler": "6.1.6",
+    "@angular/compiler-cli": "6.1.6",
+    "@angular/core": "^6.1.6",
+    "@angular/forms": "^6.1.6",
+    "@angular/http": "6.1.6",
+    "@angular/platform-browser": "6.1.6",
+    "@angular/platform-browser-dynamic": "6.1.6",
+    "@angular/router": "6.1.6",
+    "@ng-bootstrap/ng-bootstrap": "^1.1.2",
+    "@types/typeahead": "^0.11.32",
     "ajv": "^6.2.1",
     "ajv-keywords": "^3.1.0",
     "angular-popper": "^1.0.0-beta-1",
@@ -33,18 +33,21 @@
     "core-js": "^2.4.1",
     "express": "^4.16.2",
     "jquery": "^3.3.1",
+    "nodemailer": "^4.6.8",
     "popper.js": "^1.13.0",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.3.1",
+    "rxjs-compat": "^6.3.1",
     "selenium-webdriver": "^4.0.0-alpha.1",
     "ssri": "^5.2.4",
-    "typescript": "~2.5.3",
+    "typescript": "2.9.2",
     "uuid": "^3.2.1",
-    "zone.js": "^0.8.19"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.7",
-    "@angular/compiler-cli": "^5.2.3",
-    "@angular/language-service": "^5.2.3",
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "^6.1.5",
+    "@angular/compiler-cli": "6.1.6",
+    "@angular/language-service": "6.1.6",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.63",
@@ -63,7 +66,7 @@
     "protractor": "~5.3.0",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "2.9.2"
   },
   "engines": {
     "node": "8.9.4",
