@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
   selector: 'app-picks-start',
   templateUrl: './picks-start.component.html',
-  styleUrls: ['./picks-start.component.scss']
+  styleUrls: ['./picks-start.component.css']
 })
 export class PicksStartComponent implements OnInit {
 
