@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
   selector: 'app-gamespart',
   templateUrl: './gamespart.component.html',
-  styleUrls: ['./gamespart.component.scss']
+  styleUrls: ['./gamespart.component.css']
 })
 export class GamespartComponent implements OnInit {
 
